+++ Libplanet.Tests/Blockchain/NullPolicy.cs
@@ -1,3 +1,4 @@
+using System.Collections.Generic;
 using Libplanet.Action;
 using Libplanet.Blockchain;
 using Libplanet.Blockchain.Policies;
@@ -17,6 +18,8 @@ namespace Libplanet.Tests.Blockchain
             _exceptionToThrow = exceptionToThrow;
             _difficulty = difficulty;
         }
+
+        public IComparer<IBlockExcerpt> CanonicalChainComparer => new TotalDifficultyComparer();
 
         public IAction BlockAction => null;