+++ Libplanet.Tests/Blockchain/Renderers/DelayedRendererTest.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using Libplanet.Blockchain;
 using Libplanet.Blockchain.Renderers;
 using Libplanet.Blocks;
 using Libplanet.Store;
@@ -17,6 +18,7 @@ namespace Libplanet.Tests.Blockchain.Renderers
         protected static readonly IReadOnlyList<Block<DumbAction>> _chainA;
         protected static readonly IReadOnlyList<Block<DumbAction>> _chainB;
         protected static readonly Block<DumbAction> _branchpoint;
+        protected IComparer<IBlockExcerpt> _canonicalChainComparer;
         protected IStore _store;
         protected ILogger _logger;
 
@@ -67,6 +69,8 @@ namespace Libplanet.Tests.Blockchain.Renderers
                 _chainB
             );
 
+            _canonicalChainComparer = new TotalDifficultyComparer();
+
             _store = new DefaultStore(null);
             foreach (Block<DumbAction> b in _chainA.Concat(_chainB))
             {
@@ -82,6 +86,7 @@ namespace Libplanet.Tests.Blockchain.Renderers
             ArgumentOutOfRangeException e = Assert.Throws<ArgumentOutOfRangeException>(() =>
                 new DelayedRenderer<DumbAction>(
                     new AnonymousRenderer<DumbAction>(),
+                    _canonicalChainComparer,
                     _store,
                     confirmations: invalidConfirmations
                 )
@@ -103,7 +108,12 @@ namespace Libplanet.Tests.Blockchain.Renderers
                     reorgs++;
                 },
             };
-            var renderer = new DelayedRenderer<DumbAction>(innerRenderer, _store, 3);
+            var renderer = new DelayedRenderer<DumbAction>(
+                innerRenderer,
+                _canonicalChainComparer,
+                _store,
+                confirmations: 3
+            );
             Assert.Null(renderer.Tip);
             Assert.Empty(blockLogs);
             Assert.Equal(0U, reorgs);
@@ -156,7 +166,12 @@ namespace Libplanet.Tests.Blockchain.Renderers
                 BlockRenderer = (oldTip, newTip) => blockLogs.Add((oldTip, newTip)),
                 ReorgRenderer = (oldTip, newTip, bp) => reorgLogs.Add((oldTip, newTip, bp)),
             };
-            var delayedRenderer = new DelayedRenderer<DumbAction>(innerRenderer, _store, 3);
+            var delayedRenderer = new DelayedRenderer<DumbAction>(
+                innerRenderer,
+                _canonicalChainComparer,
+                _store,
+                confirmations: 3
+            );
             var renderer = new LoggedRenderer<DumbAction>(
                 delayedRenderer,
                 _logger,